requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
python-dotenv
ipykernel
seaborn
dvc
dvc-gdrive
xgboost
fastapi[standard]
python-multipart
